requirements.txt
flask
watson-developer-cloud
python-dotenv
